refactor: move url.ts to hopp.ts
