+++ packages/hoppscotch-app/helpers/import-export/import/hopp.ts
@@ -1,11 +1,18 @@
-import { pipe } from "fp-ts/function"
+import { pipe, flow } from "fp-ts/function"
 import * as TE from "fp-ts/TaskEither"
-import * as E from "fp-ts/Either"
-import { translateToNewRESTCollection } from "@hoppscotch/data"
+import * as O from "fp-ts/Option"
+import * as RA from "fp-ts/ReadonlyArray"
+import {
+  translateToNewRESTCollection,
+  HoppCollection,
+  HoppRESTRequest,
+} from "@hoppscotch/data"
+import _isPlainObject from "lodash/isPlainObject"
 import { step } from "../steps"
 import { defineImporter, IMPORTER_INVALID_FILE_FORMAT } from "."
 
 export default defineImporter({
+  id: "hoppscotch",
   name: "import.from_json",
   icon: "folder-plus",
   steps: [
@@ -19,16 +26,49 @@ export default defineImporter({
   ] as const,
   importer: ([content]) =>
     pipe(
-      E.tryCatch(
-        () => {
-          const x = JSON.parse(content)
-
-          return Array.isArray(x)
-            ? x.map((coll: any) => translateToNewRESTCollection(coll))
-            : [translateToNewRESTCollection(x)]
-        },
-        () => IMPORTER_INVALID_FILE_FORMAT
+      O.tryCatch(() => JSON.parse(content)),
+      O.chain(
+        flow(
+          makeCollectionsArray,
+          O.traverseArray(
+            flow(
+              validateCollectionSchema,
+              O.chain((collection) =>
+                O.tryCatch(() => pipe(collection, translateToNewRESTCollection))
+              )
+            )
+          ),
+          O.map(RA.toArray)
+        )
       ),
-      TE.fromEither
+      TE.fromOption(() => IMPORTER_INVALID_FILE_FORMAT)
     ),
 })
+
+/**
+ * checks if a value is a plain object
+ */
+const isPlainObject = (value: any): value is object => _isPlainObject(value)
+
+/**
+ * checks if a collection matches the schema for a hoppscotch collection.
+ * as of now we are only checking if the collection has a "v" key in it.
+ */
+const isValidCollection = (
+  collection: unknown
+): collection is HoppCollection<HoppRESTRequest> =>
+  isPlainObject(collection) && "v" in collection
+
+/**
+ * validates the collection and wraps it in an Option
+ */
+const validateCollectionSchema = (
+  collection: unknown
+): O.Option<HoppCollection<HoppRESTRequest>> =>
+  isValidCollection(collection) ? O.some(collection) : O.none
+
+/**
+ * convert single collection object into an array so it can be handled the same as multiple collections
+ */
+const makeCollectionsArray = (collections: unknown | unknown[]): unknown[] =>
+  Array.isArray(collections) ? collections : [collections]
